Fix flaky test due to slow worker startup (#232)
When creating the extension, a diskquota worker will be started
to mornitor the db. However, it might take longer than expected
until it checks whether it is ready to monitor, which will make
some tests flaky.

This patch increases the time to wait for worker startup to avoid
flaky tests.