Remove free writer gang in launcher (#233)
As launcher and worker both run in bgworker, they are clients when
connecting to segments. They know when should close the connections.
So we directly call `DisconnectAndDestroyAllGangs` to close them and
remove the code that only closes free read gangs based on timeout.

For the launcher, we destroy gangs after init_datbase_list or after each
time processing create/drop diskquota extension.

For workers, we only destroy gangs after the extension is paused.